# OrigenX/BackEnd/openapi.yaml
openapi: 3.0.3
info:
  title: The Artisanal Connection API
  description: |
    API REST para The Artisanal Connection. Conecta productores de cafe colombiano con consumidores finales.

    ## Autenticacion
    La mayoria de endpoints requieren un Firebase ID Token en el header:
    `Authorization: Bearer <firebase_id_token>`

    ## Roles
    - **CONSUMER**: consumidor final que navega el catalogo y realiza pedidos.
    - **PRODUCER**: productor de cafe que gestiona productos y ventas.
    - **ADMIN**: administrador de la plataforma.

    ## Formato de errores
    Todos los errores siguen el formato uniforme:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Descripcion del error",
        "fields": [{"field": "email", "message": "El correo ya esta registrado"}]
      }
    }
    ```
  version: 0.1.0
  contact:
    name: The Artisanal Connection
servers:
  - url: http://localhost:8000
    description: Servidor de desarrollo local
  - url: https://api.marketplacecafetero.com
    description: Servidor de produccion
...
security:
  - BearerAuth: []
paths:
...
  /api/products:
    get:
      tags: [products]
      summary: Catalogo de productos
      description: Retorna el catalogo de productos activos con paginacion y filtros opcionales. Endpoint publico. Requerimientos 9.1-9.4, 10.1-10.5.
      operationId: getCatalog
      security: []
      parameters:
        - name: page
          in: query
          description: Numero de pagina (inicia en 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Productos por pagina (maximo 20)
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 20
        - name: q
          in: query
          description: Busqueda por nombre o descripcion
          schema:
            type: string
        - name: minPrice
          in: query
          description: Precio minimo
          schema:
            type: number
            minimum: 0
        - name: maxPrice
          in: query
          description: Precio maximo
          schema:
            type: number
            minimum: 0
        - name: region
          in: query
          description: Region del productor
          schema:
            type: string
      responses:
        "200":
          description: Catalogo de productos activos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProductsResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [products]
      summary: Crear producto
      description: Crea un nuevo producto asociado al Productor autenticado. Estado inicial inactive. Requerimientos 5.1-5.5.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
            example:
              name: Cafe Especial Huila
              description: Cafe de origen unico del departamento del Huila, notas a chocolate y frutas rojas.
              price: 45000
      responses:
        "201":
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/products/{product_id}:
    get:
      tags: [products]
      summary: Detalle de producto
      description: Retorna el detalle completo de un producto. Endpoint publico. Requerimientos 11.1-11.3.
      operationId: getProductDetail
      security: []
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Detalle del producto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetailResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [products]
      summary: Actualizar producto
      description: Actualiza los campos de un producto. Solo el Productor dueno puede actualizarlo. Requerimientos 6.1-6.4.
      operationId: updateProduct
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/producers/{producer_id}:
    get:
      tags: [producers]
      summary: Obtener perfil de productor
      description: Retorna el perfil publico de un productor. Accesible por cualquier usuario. Requerimientos 4.1, 4.4.
      operationId: getProducerProfile
      security: []
      parameters:
        - $ref: "#/components/parameters/ProducerId"
      responses:
        "200":
          description: Perfil del productor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ProducerProfileData"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [producers]
      summary: Actualizar perfil de productor
      description: Crea o actualiza el perfil de un productor. Solo el Productor dueno puede actualizar su propio perfil. Retorna HTTP 400 si farmName esta vacio. Retorna HTTP 403 si el usuario autenticado no es el dueno del perfil. Requerimientos 4.1-4.4.
      operationId: updateProducerProfile
      parameters:
        - $ref: "#/components/parameters/ProducerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProducerProfileCreate"
            example:
              farm_name: Finca La Esperanza
              region: Huila
              description: Productores de cafe especial desde 1985.
              contact_info: contacto@fincaesperanza.com
      responses:
        "200":
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Perfil actualizado exitosamente.
                  data:
                    $ref: "#/components/schemas/ProducerProfileData"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/cart:
    get:
      tags: [cart]
      summary: Obtener carrito
      description: Retorna el carrito del consumidor autenticado con subtotales y total calculados. Requiere rol CONSUMER. Requerimientos 12.1, 12.6.
      operationId: getCart
      responses:
        "200":
          description: Carrito del consumidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/orders:
    post:
      tags: [orders]
      summary: Crear pedido
      description: Crea un pedido a partir del carrito del consumidor. Valida carrito no vacio, direccion valida y empresa de envio. Requiere rol CONSUMER. Requerimientos 13.1-13.5.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
            example:
              address_id: addr123
              shipping_company: servientrega
      responses:
        "201":
          description: Pedido creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Pedido creado exitosamente.
                  data:
                    $ref: "#/components/schemas/OrderData"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "503":
          $ref: "#/components/responses/ServiceUnavailableError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [orders]
      summary: Historial de pedidos
      description: Retorna todos los pedidos del consumidor autenticado, ordenados por fecha descendente. Requiere rol CONSUMER. Requerimiento 15.1.
      operationId: getOrders
      responses:
        "200":
          description: Historial de pedidos del consumidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    description: Lista de pedidos ordenados por fecha de creacion descendente
                    items:
                      $ref: "#/components/schemas/OrderData"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/orders/{order_id}:
    get:
      tags: [orders]
      summary: Detalle de pedido
      description: |
        Retorna el detalle completo de un pedido del consumidor autenticado.
        Incluye productos comprados con cantidades y precios snapshot, direccion de envio, empresa de envio y estado.
        Requiere rol CONSUMER (solo puede ver sus propios pedidos).
        Requerimiento 15.2.
      operationId: getOrderDetail
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Detalle del pedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/OrderData"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/sales:
    get:
      tags: [sales]
      summary: Panel de ventas del productor
      description: Retorna los pedidos que contienen productos del productor autenticado. Incluye total acumulado del mes en curso y anterior. Soporta filtro por rango de fechas. Requiere rol PRODUCER. Requerimientos 16.1-16.4.
      operationId: getSales
      parameters:
        - name: from_date
          in: query
          description: Fecha de inicio del filtro (YYYY-MM-DD, inclusive)
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: to_date
          in: query
          description: Fecha de fin del filtro (YYYY-MM-DD, inclusive)
          schema:
            type: string
            format: date
            example: "2024-01-31"
      responses:
        "200":
          description: Panel de ventas del productor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/addresses:
    get:
      tags: [addresses]
      summary: Listar direcciones de envio
      description: Retorna todas las direcciones de envio del consumidor autenticado. Requiere rol CONSUMER. Requerimiento 18.4.
      operationId: listAddresses
      responses:
        "200":
          description: Direcciones del consumidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AddressResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [addresses]
      summary: Registrar direccion de envio
      description: "Registra una nueva direccion de envio. Campos obligatorios: street, city, department. Limite de 5 direcciones por usuario. Requiere rol CONSUMER. Requerimientos 18.1-18.3."
      operationId: createAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressCreate"
            example:
              street: "Calle 10 No 5-20"
              city: Bogota
              department: Cundinamarca
              postal_code: "110111"
      responses:
        "201":
          description: Direccion registrada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Direccion registrada exitosamente.
                  data:
                    $ref: "#/components/schemas/AddressResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/addresses/{address_id}:
    delete:
      tags: [addresses]
      summary: Eliminar direccion de envio
      description: Elimina una direccion de envio del consumidor. Requiere rol CONSUMER.
      operationId: deleteAddress
      parameters:
        - $ref: "#/components/parameters/AddressId"
      responses:
        "200":
          description: Direccion eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Direccion eliminada exitosamente.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID Token obtenido mediante Firebase Client SDK. Incluir en el header Authorization: Bearer <token>"
  parameters:
    ProductId:
      name: product_id
      in: path
      required: true
      description: Identificador unico del producto
      schema:
        type: string
        example: prod123abc
...
    ProducerId:
      name: producer_id
      in: path
      required: true
      description: Identificador unico del productor (Firebase UID)
      schema:
        type: string
        example: uid789ghi
    OrderId:
      name: order_id
      in: path
      required: true
      description: Identificador unico del pedido
      schema:
        type: string
        example: order123jkl
...
    AddressId:
      name: address_id
      in: path
      required: true
      description: Identificador unico de la direccion
      schema:
        type: string
        example: addr789pqr
  schemas:
...
    ProductCreate:
      type: object
      required: [name, description, price]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Cafe Especial Huila
        description:
          type: string
          minLength: 1
          example: Cafe de origen unico del departamento del Huila.
        price:
          type: number
          exclusiveMinimum: 0
          description: Precio en pesos colombianos. Debe ser mayor que cero.
          example: 45000
    ProductUpdate:
      type: object
      required: [name, description, price]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Cafe Especial Huila Premium
        description:
          type: string
          minLength: 1
          example: Cafe de origen unico del departamento del Huila, cosecha 2024.
        price:
          type: number
          exclusiveMinimum: 0
          example: 48000
...
    ProductDetailResponse:
      type: object
      properties:
        message:
          type: string
          example: Producto creado exitosamente.
        data:
          type: object
          properties:
            id:
              type: string
              example: prod123abc
            producer_id:
              type: string
              example: uid789ghi
            producer_name:
              type: string
              example: Finca La Esperanza
            name:
              type: string
              example: Cafe Especial Huila
            description:
              type: string
              example: Cafe de origen unico del departamento del Huila.
            price:
              type: number
              example: 45000
            status:
              type: string
              enum: [active, inactive]
              example: inactive
            created_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00"
            updated_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00"
            images:
              type: array
              items:
                $ref: "#/components/schemas/ProductImageResponse"
...
    PaginatedProductsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductListItem"
        total:
          type: integer
          example: 45
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        has_next:
          type: boolean
          example: true
...
    ProducerProfileCreate:
      type: object
      required: [farm_name]
      properties:
        farm_name:
          type: string
          minLength: 1
          description: Nombre de la finca (obligatorio)
          example: Finca La Esperanza
        region:
          type: string
          nullable: true
          example: Huila
        description:
          type: string
          nullable: true
          example: Productores de cafe especial desde 1985.
        contact_info:
          type: string
          nullable: true
          example: contacto@fincaesperanza.com
...
    ProducerProfileData:
      type: object
      description: Datos del perfil de un productor
      properties:
        user_id:
          type: string
          example: uid789ghi
        farm_name:
          type: string
          example: Finca La Esperanza
        region:
          type: string
          nullable: true
          example: Huila
        description:
          type: string
          nullable: true
          example: Productores de cafe especial desde 1985.
        contact_info:
          type: string
          nullable: true
          example: contacto@fincaesperanza.com
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
...
    CartResponse:
      type: object
      description: Respuesta del carrito envuelta en data
      properties:
        data:
          $ref: "#/components/schemas/CartData"
...
    OrderCreate:
      type: object
      required: [address_id, shipping_company]
      properties:
        address_id:
          type: string
          description: ID de la direccion de envio del consumidor
          example: addr789pqr
        shipping_company:
          type: string
          description: ID de la empresa de envio seleccionada
          example: servientrega
...
    OrderData:
      type: object
      description: Datos completos de un pedido
      properties:
        id:
          type: string
          example: order123jkl
        consumer_id:
          type: string
          example: abc123uid
        address_snapshot:
          $ref: "#/components/schemas/AddressSnapshot"
        shipping_company:
          type: string
          example: servientrega
        shipping_cost:
          type: number
          example: 15000
        total:
          type: number
          example: 95000
        status:
          type: string
          enum: [pendiente, pagado, en_preparacion, enviado, entregado, cancelado]
          example: pendiente
        transaction_id:
          type: string
          nullable: true
          example: SIM-a1b2c3d4-e5f6-7890-abcd-ef1234567890
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemSnapshot"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
...
    SalesResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            orders:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    example: order123jkl
                  created_at:
                    type: string
                    format: date-time
                    example: "2024-01-15T10:30:00"
                  status:
                    type: string
                    enum: [pendiente, pagado, en_preparacion, enviado, entregado, cancelado]
                    example: pagado
                  shipping_company:
                    type: string
                    example: servientrega
                  items:
                    type: array
                    description: Solo los items del productor autenticado
                    items:
                      $ref: "#/components/schemas/OrderItemSnapshot"
            current_month_total:
              type: number
              description: Total acumulado de ventas del mes en curso
              example: 450000
            previous_month_total:
              type: number
              description: Total acumulado de ventas del mes anterior
              example: 380000
    AddressCreate:
      type: object
      required: [street, city, department]
      properties:
        street:
          type: string
          minLength: 1
          description: Calle (obligatorio)
          example: Calle 10 # 5-20
        city:
          type: string
          minLength: 1
          description: Ciudad (obligatorio)
          example: Bogota
        department:
          type: string
          minLength: 1
          description: Departamento (obligatorio)
          example: Cundinamarca
        postal_code:
          type: string
          nullable: true
          example: "110111"
    AddressResponse:
      type: object
      properties:
        id:
          type: string
          example: addr789pqr
        user_id:
          type: string
          example: abc123uid
        street:
          type: string
          example: Calle 10 # 5-20
        city:
          type: string
          example: Bogota
        department:
          type: string
          example: Cundinamarca
        postal_code:
          type: string
          nullable: true
          example: "110111"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
  responses:
    ValidationError:
      description: Datos de entrada invalidos (HTTP 400)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Los datos de entrada son invalidos.
              fields:
                - field: email
                  message: El correo electronico no tiene un formato valido.
                - field: password
                  message: La contrasena debe tener al menos 8 caracteres.
    UnauthorizedError:
      description: No autenticado - JWT ausente o invalido (HTTP 401)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Token de autenticacion ausente o invalido.
    ForbiddenError:
      description: No autorizado - rol insuficiente o recurso ajeno (HTTP 403)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: No tienes permiso para realizar esta accion.
    NotFoundError:
      description: Recurso no encontrado (HTTP 404)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: El recurso solicitado no fue encontrado.
...
    InternalError:
      description: Error interno del servidor (HTTP 500)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: Error interno del servidor. Por favor intenta de nuevo.
    ServiceUnavailableError:
      description: Servicio externo no disponible (HTTP 503)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: SERVICE_UNAVAILABLE
              message: El servicio de envio no esta disponible temporalmente.